py/oidc-demo/requirements.txt
Authlib==1.3.2
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
eywa-client==0.2.4
